모든기기 로그아웃 구현

+++ src/main/java/Prochat_project/repository/AlarmRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlarmRepository extends JpaRepository<AlarmEntity, Integer> {
+public interface AlarmRepository extends JpaRepository<AlarmEntity, Long> {
 
     Page<AlarmEntity> findAllByMembersId(Long MembersId, Pageable pageable);
